requirements.txt
pandas==1.4.2
dask==2022.6.0
click==8.1.3
spacy==3.3.1
fastapi==0.78.0
python-dotenv==0.20.0
transformers==4.20.0
tensorflow==2.9.1
uvicorn==0.17.6
python-multipart==0.0.5